Started to break up routes file (note that no tests change - this is a refactor, passing the same tests with better code).
